Fix regularization of cylindrical domains.
When using -cylinderfacet with a high enough value, the resulting edges
became smaller than sel (used in regularization), leading to a crash.  A
test on whether the edge to delete is a domain edge has been added. In
such a case, the deletion is now skipped. The fix applies to any domains
(not just cylinders).