Add labels to the domain faces in -T.
This bumps the tess file format to 2.0.  The labels are set to 'x0',
'x1', 'y0', 'y1', 'z0' and 'z1' for a cuboidal domain, 'z0', 'z1', 'f0',
'f1', etc. for a cylindrical domain and 'f1', 'f2', etc. for other
domains.  The labels are written in the tess file and used by -M to
label the node sets. The labels of the domain edge and vertex node sets
are built from the domain face node sets, by concatenation.