Fix nset output for meshing of voxel data in -M.
This includes the new functions 'nem_nsets_2d_hex', 'nem_nsets_1d_hex',
'nem_nsets_2dbody_hex' and 'nem_nsets_1dbody_hex', which are specific to
hex meshes and do not take any topological information as input.  Such
information is derived from the hex mesh generation. The generated nsets
are the ones of the cuboidal domain (x0, etc.) even if the domain does
not have a cuboidal shape (e.g. cylindrical). The reason for this is
that it is difficult to get relevant topological information from an hex
mesh (due to the voxelization effect).