Modify VOX for marching cube.
VoxPoly is now indexed 1 instead of 0 and allocated [0...size + 1] in
the 3 dimensions in order to get a 1-pixel layer around the volume.  The
1-pixel layer is expected to help detecting the domain faces.  Right
now, the values of VoxPoly for the 1-pixel layer are set by considering
that the domain is a cube. VoxPoly is set to -1 at x min, -2 at x max,
-3 at y min, -4 at y max, -5 at z min and -6 at zmax.  Such a definition
is consistent with what is done with the tessellations in -T.